backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain==0.0.340
langgraph==0.0.21
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
azure-identity==1.14.0
azure-storage-blob==12.18.3
langchain-community==0.0.13
msal==1.24.1
requests==2.31.0
httpx==0.25.0
pymongo[srv]==4.5.0
motor==3.3.1
sqlalchemy==2.0.21
langchain-openai==0.0.2
qdrant-client==1.6.4
python-jose==3.3.0
PyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
email-validator==2.0.0 